changing implicit hierarchy (david's idea)
